requirements.txt
pyyaml==5.4.1
tensorboard==2.4.1
rpy2==3.4.4
